Extract all paths methods in their own file
