+++ lib/data-loader.js
@@ -2,10 +2,11 @@
 
 const {StateController, Logger} = require('kite-installer');
 const {
-  hoverPath, openDocumentationInWebURL, promisifyRequest,
-  promisifyReadResponse, head, valueReportPath, membersPath,
-  usagesPath, usagePath, examplePath, parseJSON,
-} = require('./utils');
+  hoverPath, openDocumentationInWebURL, valueReportPath, membersPath,
+  usagesPath, usagePath, examplePath,
+} = require('./urls');
+const {promisifyRequest, promisifyReadResponse, head, parseJSON} = require('./utils');
+
 const {MAX_FILE_SIZE} = require('./constants');
 const {symbolId} = require('./kite-data-utils');
 const VirtualCursor = require('./virtual-cursor');
