+++ lib/elements/kite-curated-example.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const _ = require('underscore-plus');
-const {last, head, openExampleInWebURL} = require('../utils');
+const {last, head} = require('../utils');
+const {openExampleInWebURL} = require('../urls');
 const {section, renderExample} = require('./html-utils');
 
 const escapeHTML = str =>
