+++ lib/elements/kite-expand-function.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const {openDocumentationInWebURL} = require('../utils');
+const {openDocumentationInWebURL} = require('../urls');
 const {renderValueHeader, renderExtend, renderUsages, renderParameters, renderExamples, renderDefinition, renderInvocations, valueDescription} = require('././html-utils');
