+++ lib/elements/kite-expand-instance.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const {openDocumentationInWebURL} = require('../utils');
+const {openDocumentationInWebURL} = require('../urls');
 const {renderValueHeader, renderExtend, renderDefinition, renderUsages, valueDescription, renderExamples} = require('./html-utils');
 
 class KiteExpandInstance extends HTMLElement {
