+++ lib/elements/kite-expand-module.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const {openDocumentationInWebURL} = require('../utils');
+const {openDocumentationInWebURL} = require('../urls');
 const {renderValueHeader, renderExtend, renderExamples, renderUsages, renderDefinition, renderMembers, valueDescription} = require('./html-utils');
 
 class KiteExpandModule extends HTMLElement {
